Updated compilation
Added the small result for compatibility for orthogonal polynomials.
This needs to be implemented in order to do some experiments and get
further results.

Also the "generalized power basis" were fixed in the document and the
code added the Bessel basis with the corresponding compatibilities with
the division by "x".