Implemented compatibility for Orthogonal Basis
The derivation is always "almost" compatible with orthogonal
polynomials. Not only in the case of Laguerre and Hermite polynomials
(where the compatibility is complete) but also for all other families.

Following the ideas in Salvy and Benoit, we can compute a recurrence
equation R for any differential equation L such that L.y = 0 iff the
expansion of y in the corresponding basis satisfies the recurrence given
by R.

This is achieved multiplying the original equation by a particular
polynomial and computing then the (now that it is compatible) the
corresponding recurrence equation.

This implementation raise the same results as obtained in the paper by
Salvy and Benoit (although this need to be fully checked) after clearing
denominators.