Petkovsek Updated version of the code
Added a class for Scalar product of factorial basis.
This class is more flexible and relies on the original basis to obtain teh desired values.

Changed the behavior of compatible_division. Now it detects the sections in the compatibility.

Added a class for Sieved basis. Added the extension of X and E to the Sieved basis.

Removed the code inside ProductBasis. Now this code will be more generic in SievedBasis.

Updated the tests and added as tests the Apery z(2) and z(3)