Extended the behavior of compatibility
We moved the method extended_quo_rem to the main PSBasis class

Now the method compatibility detects if the input is an operator or symbolic expression and tries to extend the compatibility of the generators.

Now the method compatibility extends the set of compatible operators that we get from "compatible operators".