Version 0.3 (#1)
Completed version of patch v0.3

    Changed the CI file: now it is put together into one main file that runs all the CI actions together.
    Changes in Makefile induced by changes in Sagemath: now instead of having things installed in $SAGE_ROOT/local/bin, now it is in $SAGE_ROOT/venv/bin
    Great update in README
    Added environment.yml for use of Binder by conda.
    Updated documentation: improved links to python documentation and updated files and contents
    Added results and notebooks with experiments and examples in folder "notebooks"
    Added script to get information from OEIS
    Added several minor modules for future upgrades and generalization of the package: sequences, io, ore
    Updated setup file: removed sphinx from requirements

Remaining things to consider:
The sage version for test is older: 9,3.beta8. This implies the tests fail for them for a different representation of rational functions. This is due to the fact that we need to install git in docker container. The current version of SageMath (9.6) have a container based on Ubuntu 20.10 which is discontinued and do not allow run command sudo apt update.

This will be added as an issue on this repository to remember the reason and consequences of this problem