Removed the need to specify the endomorphisms or derivations
Previously, we save no information about the type of operators that
were compatible with a PSBasis. now there is an option in ``set_compatibility``
to define the type of the operator. We save a list of endomorphisms
and derivations.
Hence, ProductBasis and SievedBasis will use these lists to extend the
compatibility with endomorphisms and derivations.

Completed the transformation to sequences.

Now all classes have into account the changes to sequences, having the
argument base to define the basic ring of coefficients, and inheriting from
the class Map between natural numbers and implementing all the
attributes as sequences