Added a new Sequence class
The ExpressionSequences are based on symbolic expressions. These sequences can have as many variables as desired and can be combined arithmetically keeping the structure.

Added the output NaN when an error occurs evaluating a sequence (besides the oo when dividing by zero).

Restored the CI file to install pylint by default.